requirements.txt

pandas
python-telegram-bot==20.3
openpyxl
